src: lib: lore: Make `download_series` download cover letter by default
The function `download_series` that uses the b4 tool to download an
applicable mbox file of a given patchset never downloads the cover
letter. As we want to be able to visualize patch series in the future
(including the cover letter), make the download of the cover not only
possible, but also default.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>