src: plugins: kernel_install: utils: add regex support for `kw deploy --uninstall`
Currently, `kw deploy --uninstall` just supports a comma-separated
kernel list, but for some cases it's better to provide a regex pattern
that will match the desired kernels to be uninstalled. In this sense,
add support for regex arguments for the option. Note that regex
arguments should be prefixed with 'regex:'.

Closes: #344

Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>
Co-authored-by: Briza Mel Dias de Sousa <brizamel.dias@gmail.com>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>