database: kwdb: Add initial files for kw manager
To support the new `kw manage-contacts` feature, we are adding initial files and
creating database tables to store the necessary information. This includes SQL
commands to create the tables `email-contact`, `email_group` &
`email_contact_group` and the file src/manage-contacts with the initial
versions of `parse_manage_contacts_options` and the `manage_contacts_main`.

Signed-off-by: JGBSouza <joaosouzaaa12@gmail.com>