src: kw_manage_contacts: Add kw manager function add contacts
To support the new `kw manage_contacts` feature, we need a method to add contacts
to groups. In this sense, add the new functionality `--group-add` that
recieves a group and a list of names and emails of the contacts and add
the contacts and the group association in the database.

Signed-off-by: JGBSouza <joaosouzaaa12@gmail.com>