src: kw_manage_contact: add remove-email function
To support the new feature `kw manage_contacts`, it should be able to
remove contacts fom the groups based on the contact email. In this
sense, add the new functionality `group-remove-email` that recieves the
group wich the contact is member and the contact email and remove it's
association. Otherwise, if the contact has no groups after removing the
conection, also remove the contact from the database.

Signed-off-by: JGBSouza <joaosouzaaa12@gmail.com>