src: patch_track: Add initial files for kw patch-track
To support the new `kw patch-track` feature, we are adding initial files and
creating database tables to store the necessary information. This includes SQL
commands to create the table `patch`, the file src/patch with the initial
versions of `parse_patch_track_options` and the `patch_track_main` and the
initial documentation `documentation/man/features/kw-patch-track` file.

Signed-off-by: JGBSouza <joaosouzaaa12@usp.br>